Add long search bar in big header
